+++ libs/constants/release_notes.md
@@ -1,5 +1,10 @@
 # ftrack Constants library release Notes
 
+## upcoming
+
+* [new] framework/event; FTRACK_ACTION_DISCOVER_TOPIC and FTRACK_ACTION_LAUNCH_TOPIC added.
+
+
 ## v2.0.0
 2024-02-12